+++ sample04/Koopa.h
@@ -7,8 +7,8 @@ public:
 	DWORD last_time;
 	Koopa();
 	Koopa(float _x, float _y, float _cameraX, float _cameraY, int _ID, CSprite* _sprite);
-	void Move(int t);
-	void Update(int t);
+	void Move(float t);
+	void Update(float t);
 	void Render();
 	void CollisionTemp(BaseObject *obj);
 	~Koopa();
